backend-python/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
scikit-learn==1.3.0
numpy
python-dotenv
openai   # optional (only needed if you want to use OpenAI)
